# JDK-8299415: Saturate the async log buffer more often