--- a/src/hotspot/share/logging/logAsyncWriter.hpp
+++ b/src/hotspot/share/logging/logAsyncWriter.hpp
@@ -112,6 +112,11 @@ class AsyncLogWriter : public NonJavaThread {
     const size_t _capacity;
 
    public:
+
+    // Ratio of buffer that is filled in range [0,1]
+    double ratio_used() {
+      return (double)_pos / (double)_capacity;
+    }
     Buffer(size_t capacity) :  _pos(0), _capacity(capacity) {
       _buf = NEW_C_HEAP_ARRAY(char, capacity, mtLogging);
       assert(capacity >= Message::calc_size(0), "capcity must be great a token size");