@jdksjolen 2022-12-30
>also believe that less uses of notify() means that the performance implication of using asynchronous logging is improved. Running stat perf consistently shows a smaller amount of context switches when run with this change, which supports this hypothesis.

This is probably fairly weak proof of this, and a day later I'm not so sure that this is true :-). Still, the change is good imho!