@navyxliu 2022-12-30
I think your change of fraction_used() is reasonable. We don't need to notify the flush thread in very early stage. As long as 0.3 doesn't cause overflow for `-Xlog:all=trace`, it's good for me. Your histogram and ctx switch counter manifests its effectiveness.