@jdksjolen 2022-12-31
>Does this increase context switch? In particular, hotspot may not enable unified log at all. The flush thread wakes up every second for nothing?

Yes, it increases context switching. If Hotspot hasn't enabled `-Xlog:async` then this code won't run.

>I understand that we somehow need to flush the queue if the log stream is end and fraction_used() is still lower than 0.3. I think we can remember a historic fraction_used. As long as this value increases, it suggests that logs still trickle in. we don't need to worry about it if so.

The reasoning for the timeout was two fold:

1. As you said: the log stream is end and fraction_used() is still lower than 0.3, we should still log in a timely manner.
2. Flushing logs in a timely manner, making it suitable for interactive use

Let's say that a historic fraction is used, then where would the comparison be placed? The only place I can think of is the `enqueue` method, but then we already know that messages are being supplied. We could compute a rate of change and if that rate is too low then we flush each time.  Both of these solutions depend on an infinite stream of logs (however far apart the messages are) if the timeout is not used. Perhaps this should be combined with a much longer timeout? Such as 30 seconds. I think spuriously waking up every 30 seconds is acceptable.

>Using this algorithm, I think we can make even bigger threshold. The reasonable threshold is up to the speed of IO. In reality, the speed of IO is not known until runtime. I think we can make it adaptive in the future. if we detect overflow, we lower the threshold down to 0.0.

It's reasonable for this RFE to lower the threshold directly down to 0.0 if overflow is detected, I agree that it's a good idea for a future RFE to implement something more dynamic.