@navyxliu 2023-01-02
hi, @jdksjolen 

> Yes, it increases context switching. If Hotspot hasn't enabled -Xlog:async then this code won't run.

Fair enough. One corner case is that users declare '-Xlog:async' but don't provide any '-Xlog:' selections. They may postpone to select logs in runtime using dcmd. We can ignore that because the compromise is acceptable. 

> Both of these solutions depend on an infinite stream of logs (however far apart the messages are) if the timeout is not used. Perhaps this should be combined with a much longer timeout? Such as 30 seconds. I think spuriously waking up every 30 seconds is acceptable.

I agree. No matter what, we can't put the flush thread in sleep without a timeout argument. We can't flush holding messages until termination. The fixed timeout 1s isn't perfect but I think it's reasonable. Your patch looks good to me.