@dholmes-ora 2023-01-03
There is obviously a spectrum of changes possible based on how big a buffer is used and how often it is flushed - can the user already control these two things? If yes then I don't see a justification to change the current behaviour. If no then surely that is what is needed?