@navyxliu 2023-01-03
hi, @dholmes-ora 

Current implementation lacks the parameter of "how often it is flushed". Logsites notify the flush thread for every single message and it is the most conservative way. The general idea is to flush as many messages as possible each time. This amortizes the cost of synchronization and external IO. I think this patch is toward this goal.