@dholmes-ora 2023-01-04
The premise of this patch was "the buffer is being under-utilised", so it delays flushing so that the buffer gets more full. This seems an arbitrary selection - we could have made the buffer smaller instead. I'm not seeing a rationale for the current change that justifies changing the behaviour for all users in a way that they cannot compensate for. If buffer utilisation is a metric of interest (and I'm not convinced it is) then we should be providing a way to control the buffer size (already there) and the flush rate.