@jdksjolen 2023-01-05
> 

Hi David, I was surprised to hear that your intuition regarding this went against my own. So, that caused me to re-do the measurements and I realised I made a big mistake: I ran the tests without optimisations (slowdebug). I then decided to re-run the measurements with an optimised build (no `perf stat` as this changes the performance drastically).

The results of those measurements showed that `all=trace` do cause messages to be dropped with my patch. `all=debug` shows behavior similar to my initial testing: my patch fully saturates the buffer while original code doesn't. I suspect that there is huge lock contention on the `AsyncLogLocker` in the `all=trace` case, which makes it very difficult for the writer to gain access in time for the buffer to be filled and messages to be dropped.

With that in mind, maybe what should be looked into isn't this patch, but rather replacing the buffer with a lock-free circular one (for example, maybe there are better fits). That is, if we want better performance.

@navyxliu , maybe we should avoid merging this patch as it leads to increased complexity while there are other measures that could give much bigger gains?