@navyxliu 2023-01-11
hi, @jdksjolen 
It's fine to drop it. We should avoid premature optimizations. I think the number of context switch is a good metric. "buffer utilisation" may not be a primary metric.  

I still have a task in my backlog. I plan to compress `Message` when consecutive messages are from the same LogOutput. I can save a pointer(8bytes) for each message.

Yes, lock-free circular is better, but I am not sure we have a scenario like Linux kernel printk, which is always highly contented in kernel threads. If we ignore the extreme case such as `trace=all`, I only observed heat lock contention among current GC threads when 'gc*=debug'.

thanks,
--lx